References lp:1093063 - updated to wsrep API r92

+++ sql/log.cc
@@ -544,7 +544,7 @@ void thd_binlog_rollback_stmt(THD * thd)
   with the exception that here we write in buffer instead of log file.
  */
 
-int wsrep_write_cache(IO_CACHE *cache, uchar **buf, int *buf_len)
+int wsrep_write_cache(IO_CACHE *cache, uchar **buf, size_t *buf_len)
 {
 
   if (reinit_io_cache(cache, READ_CACHE, 0, 0, 0))
@@ -577,7 +577,7 @@ int wsrep_write_cache(IO_CACHE *cache, uchar **buf, int *buf_len)
       *buf = (uchar *)my_realloc(*buf, total_length+length, MYF(0));
       if (!*buf)
       {
-        WSREP_ERROR("io cache write problem: %d %d", *buf_len, length);
+        WSREP_ERROR("io cache write problem: %zu %d", *buf_len, length);
         return ER_ERROR_ON_WRITE;
       }
       buf_ptr = *buf+total_length;
@@ -586,7 +586,7 @@ int wsrep_write_cache(IO_CACHE *cache, uchar **buf, int *buf_len)
     {
       if (buf_ptr != NULL)
       {
-        WSREP_ERROR("io cache alloc error: %d %d", *buf_len, length);
+        WSREP_ERROR("io cache alloc error: %zu %d", *buf_len, length);
         my_free(*buf);
       }
       if (length > 0) 
@@ -5287,7 +5287,7 @@ err:
     {
       IO_CACHE* cache= get_trans_log(thd);
       uchar* buf= NULL;
-      int buf_len= 0;
+      size_t buf_len= 0;
 
       if (wsrep_emulate_bin_log)
         thd->binlog_flush_pending_rows_event(false);
